# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: "1.0.0"
  description: Sakani Dz - Real Estate & Tourism Platform API
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Vendor:
      type: object
      required: [id, firstName, lastName, phone, email, status, allowedPropertyTypes, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: string
        email:
          type: string
        code:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, renewal_required, blocked, rejected]
        allowedPropertyTypes:
          type: array
          items:
            type: string
        isVerified:
          type: boolean
        lastAccessAt:
          type: ["string", "null"]
          format: date-time
        blockedAt:
          type: ["string", "null"]
          format: date-time
        subscriptionExpiresAt:
          type: ["string", "null"]
          format: date-time
        totalViews:
          type: integer
        totalProperties:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Subscription:
      type: object
      required: [id, vendorId, amount, propertyType, paidAt, expiresAt, status, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        vendor:
          $ref: "#/components/schemas/VendorSummary"
        amount:
          type: number
        propertyType:
          type: string
          enum: [apartment, villa, site]
        paidAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, expired, cancelled]
        createdAt:
          type: string
          format: date-time
...
    VendorListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Vendor"
        pagination:
          $ref: "#/components/schemas/Pagination"
    SubscriptionListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Subscription"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    VendorRankingListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/VendorRanking"
...
    VendorPermissionsInput:
      type: object
      required: [allowedPropertyTypes]
      properties:
        allowedPropertyTypes:
          type: array
          items:
            type: string
...
    SubscriptionInput:
      type: object
      required: [vendorId, propertyType, amount]
      properties:
        vendorId:
          type: integer
        propertyType:
          type: string
          enum: [apartment, villa, site]
        amount:
          type: number
paths:
...
  /admin/vendors:
    get:
      operationId: listAdminVendors
      tags: [admin]
      summary: List all vendors (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, active, renewal_required, blocked, rejected]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorListResponse"
  /admin/vendors/{id}:
    get:
      operationId: getAdminVendor
      tags: [admin]
      summary: Get vendor detail (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /admin/vendors/{id}/approve:
    post:
      operationId: approveVendor
      tags: [admin]
      summary: Approve vendor and generate unique code
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /admin/vendors/{id}/reject:
    post:
      operationId: rejectVendor
      tags: [admin]
      summary: Reject vendor registration
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /admin/vendors/{id}/block:
    post:
      operationId: blockVendor
      tags: [admin]
      summary: Block a vendor
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /admin/vendors/{id}/reactivate:
    post:
      operationId: reactivateVendor
      tags: [admin]
      summary: Reactivate a blocked vendor
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /admin/vendors/{id}/permissions:
    patch:
      operationId: updateVendorPermissions
      tags: [admin]
      summary: Update vendor allowed property types (La Clأ©)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorPermissionsInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /admin/subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List subscriptions (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, expired, cancelled]
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionListResponse"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Record a subscription payment (admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /admin/subscriptions/{id}:
    get:
      operationId: getSubscription
      tags: [subscriptions]
      summary: Get subscription detail
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [stats]
      summary: Get admin dashboard statistics
      security:
        - BearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
            default: month
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/stats/top-vendors:
    get:
      operationId: getTopVendors
      tags: [stats]
      summary: Get top 5 vendors ranking
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorRankingListResponse"
  /admin/stats/by-wilaya:
    get:
      operationId: getStatsByWilaya
      tags: [stats]
      summary: Get property statistics per wilaya
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WilayaStatsListResponse"
  /admin/password-reset-requests:
    get:
      operationId: listPasswordResetRequests
      tags: [admin]
      summary: List pending vendor password reset requests
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetRequestListResponse"
  /admin/password-reset-requests/{id}/approve:
    post:
      operationId: approvePasswordResetRequest
      tags: [admin]
      summary: Approve a vendor password reset request
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetRequest"
  /admin/reports:
    get:
      operationId: listAdminReports
      tags: [reports]
      summary: List property reports (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, resolved, dismissed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportListResponse"
  /admin/reports/{id}/resolve:
    post:
      operationId: resolveReport
      tags: [reports]
      summary: Resolve a property report (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /admin/reports/{id}/dismiss:
    post:
      operationId: dismissReport
      tags: [reports]
      summary: Dismiss a property report (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  # â”€â”€ Public stats â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
...